- Removed some unnecessary dependencies - Dependencies are now stored in include.h - Changed UINT to size_t and added usage of auto - updated to VS2017

+++ Chroma Quickstart Example/Template.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "include.h"
 #ifndef _CHROMASDKIMPL_H_
 #define _CHROMASDKIMPL_H_
 
@@ -40,7 +41,7 @@ public:
 	BOOL Initialize();
 
 
-	void ResetEffects(UINT DeviceType);
+	void ResetEffects(size_t DeviceType);
 	//Define your methods here
 	BOOL example_keyboard();
 	BOOL example_mouse();
